+++ packages/components/organisms/f-header/src/components/Header.vue
@@ -248,10 +248,7 @@ html:global(.is-navInView) {
 
     .c-header--highlightBg {
         background-color: $color-support-brand-01;
-
-        @include media('>mid') {
-            min-height: 88px;
-        }
+        min-height: 88px;
     }
 
     .c-header-container {