+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v7.4.0
+v8.0.0
 ------------------------------
 *December 9, 2021*
 
@@ -12,6 +12,7 @@ v7.4.0
 - Make icons visible in mobile nav for highlight theme.
 
 ### Added
+- (BREAKING) `tallBelowMid` prop to support taller header for mobile viewports (currently used in highlight theme).
 - Supporting visual regression tests.